Update 2021-12-07-6DHypercubeJH-23-01-02.md
The model is SCALED so that the 4D hypercubes from which it is generated can be seen more clearly. The 4D hypercubes are represented in their "cube by cube" form (see previous posts). It can be built as shown using Zometool but the large blue struts used are now obsolete.