Update 2021-12-07-6DHypercubeJH-23-01-02.md
This is a model in vZome of a Zometool model of a SCALED 3D projection of a 6D Hypercube. It is scaled so all the 4D hypercube representations can be seen more easily. The 4D hypercubes are shown using their "cube-by-cube" representations. Unfortunately the large blue struts are now obsolete.